revert `f` existence check
`exists()` throws a false negative in the examples. This might be very confusing for users, and since I have little control over example evaluation, I simply removed this feature and reinstalled a uniform message: equal for named and anonymous functions.

By way of compromise, the parentheses after the function name in the message are now gone, as well. This makes more sense if `f` is an anonymous function (as opposed to having parentheses, which would actually be non-syntactical). It might even declutter the message in general.