rearrange output function's body elements
This eliminates the `f_name` variable from the factory-made function. It's no longer neede because the substring it used to bind is now integrated into the injected variables `msg_apply` and `class_expr`.

Also, the output function's body now starts with the conditional message. This makes the body more logical in its structure and might help with memory.